+++ source/headers/bullet.h
@@ -12,7 +12,7 @@ class Bullet {
     std::string type;
 public:
     sf::Sprite* bulletSprite;
-    Bullet(int, int, const std::string&, Path*);
+    Bullet(float, float, const std::string&, Path*);
     ~Bullet();
     void moveBullet();
     bool isActive;
